pyproject.toml
[tool.poetry]
name = "JotterJoy"
version = "0.1.0"
description = ""
authors = ["Anis <anisanissakkaf@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.12"
groq = "^0.8.0"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.20.0"
black = { version = "^23.1.0", allow-prereleases = true }
flake8 = "^6.0.0"
djlint = "^1.34.0"
autoflake = "^2.2.1"
pyinstaller = "^6.7.0"

[tool.poe.tasks]
start = "fastapi dev src/jotterjoy/app/main.py"
build = "python src/jotterjoy/app/pyinstaller.py"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core", "cython"]
build-backend = "poetry.core.masonry.api"
